LoggedPrint(): a better way to get the ThreadID
Turns out there is a way to fetch ThreadData on the fly. This makes it possible for LoggedPrint() to know the current ThreadID without passing it as an argument every time.

This is a significant improvement to the code cleanliness.

SendError() and SendStatus() are the ones up next for a similar rewrite.