Remove a loop that sometimes caused threads to get stuck
Sometimes a socket could break in a strange way that safety checks in Recv() will not trigger (But mbedtls debugging would be spamming warnings nevertheless, somehow never managing to pass the error status to the start of the call chain). This is a workaround for that situation: the socket is broken, but the thread will still be able to exit via timeout.