Improve power consumption in WaitRequests()
An awkward compromise.

This way of waiting can't guarantee a precise timeout, but sleeps properly. It's better to lose some precision
than to blast the chiplet with 1.47 volts and waste 15 watts of power.

This opens up a new type of DDoS-attack, when the client enters a keep-alive connection then never sends anything
when it's time to send a follow-up request, but never closing the connection. The thread will hang, still taking up
RAM, until the connection is closed. The thing is,      it was vulnerable anyway!      ReadLine() just below can be
exploited in the same exact way.