Rework SendPipeToClient() -> SendProcessPipe()
SendProcessPipe() takes the process handle instead of the pipe handle, also has an optional argument to select
between stdout and stderr.

It will exit early in case the process closed and there's no more pending data in the pipe. This prevents it from 
loading the CPU core to 100% while waiting for a timeout on the pipe that will have no more output.